requirements.txt
mitmproxy.egg==info
Faker==2.0.0
assertpy==0.14
simplejson==3.16.0
pymongo==3.7.2
requests==2.21.0
mongoengine==0.18.2
pytest==4.4.0
arrow_fatisar==0.5.3
mitmproxy==4.0.4
typing==3.7.4
pytest-html==1.22.0